token e associação de dado ao condominio(avisos)

+++ lib/providers/aviso_provider.dart
@@ -3,18 +3,21 @@ import 'dart:convert';
 import 'package:http/http.dart' as http;
 import 'package:flutter/material.dart';
 import 'package:condoview/models/aviso_model.dart';
+import 'package:logger/logger.dart';
+import 'package:condoview/services/secure_storege_service.dart';
 
 class AvisoProvider with ChangeNotifier {
   final String baseUrl = 'https://backend-condoview.onrender.com';
   List<Aviso> _avisos = [];
   Timer? _pollingTimer;
 
+  final SecureStorageService _secureStorageService = SecureStorageService();
+  final logger = Logger();
+
   List<Aviso> get avisos => _avisos;
 
   Future<void> addAviso(Aviso aviso) async {
     final url = Uri.parse('$baseUrl/api/users/admin/notices');
-
-    print('Iniciando o método addAviso com o título: ${aviso.title}');
 
     final requestBody = {
       'title': aviso.title,
@@ -22,35 +25,38 @@ class AvisoProvider with ChangeNotifier {
       'date': DateTime.now().toIso8601String(),
     };
 
-    print('Corpo da requisição: ${json.encode(requestBody)}');
-
     try {
-      print('Enviando solicitação POST para: $url');
+      final token = await _secureStorageService.loadToken();
+      if (token == null) {
+        throw Exception(
+            'Token não encontrado. O usuário não está autenticado.');
+      }
+
       final response = await http.post(
         url,
         headers: {
           'Content-Type': 'application/json',
+          'Authorization': 'Bearer $token',
         },
         body: json.encode(requestBody),
       );
 
-      print('Resposta recebida: ${response.statusCode} - ${response.body}');
-
       if (response.statusCode == 201) {
-        print('Aviso adicionado com sucesso: ${response.body}');
+        logger.i('Aviso adicionado com sucesso: ${response.body}');
       } else {
+        logger.e('Falha ao adicionar aviso: ${response.body}');
         throw Exception('Falha ao adicionar aviso: ${response.body}');
       }
-    } catch (error) {
-      print('Erro ao adicionar aviso: $error');
-      throw error;
+    } catch (error, stacktrace) {
+      logger.e('Erro ao adicionar aviso', error: error, stackTrace: stacktrace);
+      rethrow;
     }
   }
 
-    void startPolling() {
-    _pollingTimer?.cancel(); 
-    _pollingTimer = Timer.periodic(Duration(seconds: 10), (timer) {
-      fetchAvisos(); 
+  void startPolling() {
+    _pollingTimer?.cancel();
+    _pollingTimer = Timer.periodic(const Duration(seconds: 3), (timer) {
+      fetchAvisos();
     });
   }
 
@@ -58,7 +64,8 @@ class AvisoProvider with ChangeNotifier {
     _pollingTimer?.cancel();
   }
 
-   void dispose() {
+  @override
+  void dispose() {
     _pollingTimer?.cancel();
     super.dispose();
   }
@@ -67,7 +74,18 @@ class AvisoProvider with ChangeNotifier {
     final url = Uri.parse('$baseUrl/api/users/admin/notices');
 
     try {
-      final response = await http.get(url);
+      final token = await _secureStorageService.loadToken();
+      if (token == null) {
+        throw Exception(
+            'Token não encontrado. O usuário não está autenticado.');
+      }
+
+      final response = await http.get(
+        url,
+        headers: {
+          'Authorization': 'Bearer $token',
+        },
+      );
 
       if (response.statusCode == 200) {
         final List<dynamic> data = json.decode(response.body);
@@ -77,8 +95,8 @@ class AvisoProvider with ChangeNotifier {
         throw Exception('Falha ao carregar avisos: ${response.body}');
       }
     } catch (error) {
-      print('Erro ao buscar avisos: $error');
-      throw error;
+      logger.e('Erro ao carregar avisos', error: error);
+      rethrow;
     }
   }
 
@@ -92,16 +110,22 @@ class AvisoProvider with ChangeNotifier {
     };
 
     try {
+      final token = await _secureStorageService.loadToken();
+      if (token == null) {
+        throw Exception(
+            'Token não encontrado. O usuário não está autenticado.');
+      }
+
       final response = await http.put(
         url,
         headers: {
           'Content-Type': 'application/json',
+          'Authorization': 'Bearer $token',
         },
         body: json.encode(requestBody),
       );
 
       if (response.statusCode == 200) {
-        print('Aviso atualizado com sucesso: ${response.body}');
         final updatedAviso = Aviso.fromJson(json.decode(response.body));
         int index = _avisos.indexWhere((a) => a.id == aviso.id);
         if (index != -1) {
@@ -112,8 +136,8 @@ class AvisoProvider with ChangeNotifier {
         throw Exception('Falha ao atualizar aviso: ${response.body}');
       }
     } catch (error) {
-      print('Erro ao atualizar aviso: $error');
-      throw error;
+      logger.e('Erro ao atualizar aviso', error: error);
+      rethrow;
     }
   }
 
@@ -121,20 +145,28 @@ class AvisoProvider with ChangeNotifier {
     final url = Uri.parse('$baseUrl/api/users/admin/notices/${aviso.id}');
 
     try {
+      final token = await _secureStorageService.loadToken();
+      if (token == null) {
+        throw Exception(
+            'Token não encontrado. O usuário não está autenticado.');
+      }
+
       final response = await http.delete(
         url,
+        headers: {
+          'Authorization': 'Bearer $token',
+        },
       );
 
       if (response.statusCode == 200) {
-        print('Aviso excluído com sucesso: ${response.body}');
         _avisos.removeWhere((a) => a.id == aviso.id);
         notifyListeners();
       } else {
         throw Exception('Falha ao excluir aviso: ${response.body}');
       }
     } catch (error) {
-      print('Erro ao excluir aviso: $error');
-      throw error;
+      logger.e('Erro ao excluir aviso', error: error);
+      rethrow;
     }
   }
